vertex data is currently stored per object Visibility tests fail possibly monte-carlo integration sample point generation is wrong(the generated points are not co-planar with te respective face) color interpolation does not work Check actual radiosity value computation and form-factors The intensity per face is still monochromatic (a single float)
But overall, the code works :D

+++ CustomRadiosity/RadiosityFace.h
@@ -7,9 +7,9 @@ struct RadiosityFace
 {
 	ObjectModel* model;
 	int faceIndex;
-	glm::vec3 totalRadiosity;
-	glm::vec3 emission; // the energy emitted per unit of patch area
-	glm::vec3 unshotRadiosity;
+	glm::dvec3 totalRadiosity;
+	glm::dvec3 emission; // the energy emitted per unit of patch area
+	glm::dvec3 unshotRadiosity;
 };
 
 #endif
